add tests for path-utils.ts
